Created endpoint to update existing menu category
